requirements.txt
requests
click
setuptools
pyyaml
